Reorganize the base classes
